refactor: split user api key routes
